added to webpack.base
